Separate Class and Item creation into separate pages and separate tabs on sidebar Remove adding to stripe database to refactor payment solution
